float32 - float64

+++ models/DepositDetails.go
@@ -1,7 +1,7 @@
 package models
 
 type DepositDetails struct {
-	Amount      float32 `form:"amount"`
+	Amount      float64 `form:"amount"`
 	PhoneNumber string  `form:"phonenumber"`
 	UserName    string  `form:"username"`
 }
