+++ models/Mt4Account.go
@@ -5,8 +5,8 @@ type Mt4Account struct {
 	Broker     string  `json:"broker,omitempty"`
 	Currency   string  `json:"currency,omitempty"`
 	Server     string  `json:"server,omitempty"`
-	Balance    float32 `json:"balance"`
-	Equity     float32 `json:"equity,omitempty"`
-	Margin     float32 `json:"margin,omitempty"`
-	FreeMargin float32 `json:"freemargin,omitempty"`
+	Balance    float64 `json:"balance"`
+	Equity     float64 `json:"equity,omitempty"`
+	Margin     float64 `json:"margin,omitempty"`
+	FreeMargin float64 `json:"freemargin,omitempty"`
 }
